feature: extract media preview renderers
